修改clientMetaInfo id 类型 (#173)

+++ qmq-metaserver/src/main/java/qunar/tc/qmq/meta/store/Store.java
@@ -70,7 +70,7 @@ public interface Store {
 
     List<ClientMetaInfo> queryConsumerBySubject(String subject);
 
-    int updateTimeByIds(List<Integer> ids);
+    int updateTimeByIds(List<Long> ids);
 
     void batchUpdateMetaInfo(String clientId ,String subjects);
 }
